requirements.txt
Package                       Version
----------------------------- -------
asgiref                       3.7.2
black                         24.1.1
click                         8.1.7
colorama                      0.4.6
Django                        5.0.1
djangorestframework           3.14.0
djangorestframework-simplejwt 5.3.1
mypy-extensions               1.0.0
packaging                     23.2
pathspec                      0.12.1
pillow                        10.2.0
pip                           22.3.1
platformdirs                  4.2.0
psycopg2                      2.9.9
PyJWT                         2.8.0
python-decouple               3.8
pytz                          2024.1
ruff                          0.2.0
setuptools                    65.5.0
sqlparse                      0.4.4
tzdata                        2023.4
